TST: Explicitly use np.int32 in test matrix indices.

+++ tests/conftest.py
@@ -21,13 +21,15 @@ from scipy import sparse
 def davis_example_chol():
     """Return a small example matrix from Davis (2006)."""
     N = 11
-    rows = np.array([5, 6, 2, 7, 9, 10, 5, 9, 7, 10, 8, 9, 10, 9, 10, 10])
-    cols = np.array([0, 0, 1, 1, 2, 2, 3, 3, 4, 4, 5, 5, 6, 7, 7, 9])
+    rows = np.array(
+        [5, 6, 2, 7, 9, 10, 5, 9, 7, 10, 8, 9, 10, 9, 10, 10], dtype=np.int32
+    )
+    cols = np.array([0, 0, 1, 1, 2, 2, 3, 3, 4, 4, 5, 5, 6, 7, 7, 9], dtype=np.int32)
     # vals = np.ones(len(rows), dtype=np.float64)
     rng = np.random.default_rng(565656)
     vals = rng.random(len(rows), dtype=np.float64)
     L = sparse.coo_array((vals, (rows, cols)), shape=(N, N))
-    A = L + L.T   # make it symmetric
+    A = L + L.T  # make it symmetric
     A.setdiag(N)  # make it strongly positive definite
     return A.tocsc()
 
@@ -38,10 +40,14 @@ def davis_example_chol():
 def davis_example_qr():
     """Return a small example matrix from Davis (2006)."""
     N = 8
-    rows = np.array([0, 1, 2, 3, 4, 5, 6,
-                     3, 6, 1, 6, 0, 2, 5, 7, 4, 7, 0, 1, 3, 7, 5, 6])
-    cols = np.array([0, 1, 2, 3, 4, 5, 6,
-                     0, 1, 2, 2, 3, 3, 4, 4, 5, 5, 6, 6, 6, 6, 7, 7])
+    rows = np.array(
+        [0, 1, 2, 3, 4, 5, 6, 3, 6, 1, 6, 0, 2, 5, 7, 4, 7, 0, 1, 3, 7, 5, 6],
+        dtype=np.int32,
+    )
+    cols = np.array(
+        [0, 1, 2, 3, 4, 5, 6, 0, 1, 2, 2, 3, 3, 4, 4, 5, 5, 6, 6, 6, 6, 7, 7],
+        dtype=np.int32,
+    )
     # rng = np.random.default_rng(565656)
     # vals = rng.random(len(rows), dtype=np.float64)
     vals = np.ones(len(rows), dtype=np.float64)
